requirements.txt
streamlit==1.28.1
tensorflow==2.14.0
numpy==1.24.3
matplotlib==3.8.0
seaborn==0.13.0
Pillow==10.0.1
streamlit-drawable-canvas==0.2.2
